pyproject.toml
[tool.poetry]
name = "file-organizer"
version = "0.1.0"
description = "\"Organizes files in a directory by file type.\""
authors = ["Marouane ELBISSOURI"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click-didyoumean = "^0.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
